Module Comptabilite cours v1.0

+++ app/Domaine/Business/SapeurBusiness.php
@@ -5,6 +5,7 @@ namespace App\Domaine\Business;
 
 use App\Domaine\SPI\SapeurRepository;
 use App\Domaine\Exceptions\ArrayException;
+use App\Infrastructure\Models\Ecriture;
 use App\Infrastructure\Models\GradeSapeur;
 use App\Infrastructure\Models\Mutation;
 use App\Infrastructure\Models\Sapeur;
@@ -167,6 +168,10 @@ class SapeurBusiness
 
     public function removeCours(int $sapeurId, int $coursSapeurId)
     {
+        // Check que le cours n'est pas lié à une écriture
+        if (Ecriture::where('cours_sapeur', '=', $coursSapeurId)->count() > 0) {
+            throw new ArrayException([], 'Impossible de supprimer un cours facturé');
+        }
         $this->repository->removeCours($sapeurId, $coursSapeurId);
     }
